Dev

See merge request cfs/fndc!822

+++ service/src/main/java/ir/demisco/cfs/service/impl/DefaultFinancialDocument.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.databind.util.ISO8601Utils;
 import ir.demisco.cfs.model.dto.request.ControlFinancialAccountNatureTypeInputRequest;
 import ir.demisco.cfs.model.dto.request.FinancialDocumentSecurityInputRequest;
 import ir.demisco.cfs.model.dto.request.FinancialPeriodLedgerStatusRequest;
+import ir.demisco.cfs.model.dto.request.FinancialPeriodRequest;
 import ir.demisco.cfs.model.dto.request.FinancialPeriodStatusRequest;
 import ir.demisco.cfs.model.dto.response.FinancialCentricAccountDto;
 import ir.demisco.cfs.model.dto.response.FinancialDocumentAccountDto;
@@ -434,6 +435,13 @@ public class DefaultFinancialDocument implements FinancialDocumentService {
         if (financialPeriodStatus.getPeriodStatus() == 0L || financialPeriodStatus.getMonthStatus() == 0L) {
             throw new RuleException("دوره مالی و ماه عملیاتی سند مقصد میبایست در وضعیت باز باشند");
         }
+        FinancialPeriodLedgerStatusRequest financialPeriodLedgerStatusRequest = new FinancialPeriodLedgerStatusRequest();
+        financialPeriodLedgerStatusRequest.setFinancialDocumentId(responseFinancialDocumentStatusDto.getId());
+        FinancialPeriodStatusResponse financialPeriodStatusResponse = getFinancialPeriodStatus(financialPeriodLedgerStatusRequest);
+        if (financialPeriodStatusResponse.getPeriodStatus() == 0L || financialPeriodStatusResponse.getMonthStatus() == 0L) {
+            throw new RuleException("دوره مالی و ماه مربوط به دفتر مالی میبایست در وضعیت باز باشند");
+        }
+
         FinancialDocument financialDocument = financialDocumentRepository.findById(responseFinancialDocumentStatusDto.getId()).orElseThrow(() -> new RuleException("fin.financialDocument.notExistDocument"));
         FinancialDocumentStatus financialDocumentStatus =
                 documentStatusRepository.findFinancialDocumentStatusByCode(responseFinancialDocumentStatusDto.getFinancialDocumentStatusCode());
@@ -462,7 +470,7 @@ public class DefaultFinancialDocument implements FinancialDocumentService {
 
     private List<FinancialDocumentErrorDto> validationSetStatus(FinancialDocument financialDocument) {
         List<FinancialDocumentErrorDto> financialDocumentErrorDtoList = checkValidationSetStatus(financialDocument);
-        FinancialDocument document = financialDocumentRepository.getActivePeriodAndMontInDocument(financialDocument.getId());
+        Long document = financialDocumentRepository.getActivePeriodAndMontInDocument(financialDocument.getId());
         if (document == null) {
             FinancialDocumentErrorDto activeMountStatus = new FinancialDocumentErrorDto();
             activeMountStatus.setFinancialDocumentId(financialDocument.getId());
@@ -721,6 +729,15 @@ public class DefaultFinancialDocument implements FinancialDocumentService {
         if (financialPeriodStatus.getPeriodStatus() == 0L || financialPeriodStatus.getMonthStatus() == 0L) {
             throw new RuleException("دوره مالی و ماه عملیاتی سند مقصد میبایست در وضعیت باز باشند");
         }
+
+        FinancialPeriodLedgerStatusRequest financialPeriodLedgerStatusRequest = new FinancialPeriodLedgerStatusRequest();
+        financialPeriodLedgerStatusRequest.setFinancialDocumentId(financialDocumentDto.getId());
+        FinancialPeriodStatusResponse financialPeriodStatusResponse = getFinancialPeriodStatus(financialPeriodLedgerStatusRequest);
+        if (financialPeriodStatusResponse.getPeriodStatus() == 0L || financialPeriodStatusResponse.getMonthStatus() == 0L) {
+            throw new RuleException("دوره مالی و ماه مربوط به دفتر مالی میبایست در وضعیت باز باشند");
+        }
+
+
         entityManager.createNativeQuery(" update fndc.financial_document_item " +
                 "   set description = replace(description,:description,:newDescription) " +
                 "   where id in (:FinancialDocumentItemIdList) " +
@@ -744,17 +761,23 @@ public class DefaultFinancialDocument implements FinancialDocumentService {
         financialDocumentSecurityInputRequest.setFinancialDocumentItemId(null);
         financialDocumentSecurityInputRequest.setSecurityModelRequest(null);
         financialDocumentSecurityService.getFinancialDocumentSecurity(financialDocumentSecurityInputRequest);
-
-        FinancialDocument financialDocument = financialDocumentRepository.getActivePeriodAndMontInDocument(document.getId());
+        Long financialDocument = financialDocumentRepository.getActivePeriodAndMontInDocument(document.getId());
+        FinancialPeriodLedgerStatusRequest financialPeriodLedgerStatusRequest = new FinancialPeriodLedgerStatusRequest();
+        financialPeriodLedgerStatusRequest.setFinancialDocumentId(financialDocumentId);
+        FinancialPeriodStatusResponse financialPeriodStatusResponse = getFinancialPeriodStatus(financialPeriodLedgerStatusRequest);
+        if (financialPeriodStatusResponse.getPeriodStatus() == 0L || financialPeriodStatusResponse.getMonthStatus() == 0L) {
+            throw new RuleException("دوره مالی و ماه مربوط به دفتر مالی میبایست در وضعیت باز باشند");
+        }
         if (financialDocument == null) {
             throw new RuleException("fin.financialDocument.openStatusPeriod");
+
         } else {
             List<FinancialDocumentItem> financialDocumentItemList = financialDocumentItemRepository.findByFinancialDocumentIdAndDeletedDateIsNull(financialDocumentId);
             deleteDocumentItem(financialDocumentItemList);
             List<FinancialDocumentNumber> financialDocumentNumberList = financialDocumentNumberRepository.findByFinancialDocumentIdAndDeletedDateIsNull(financialDocumentId);
             financialDocumentNumberList.forEach(financialDocumentNumber -> financialDocumentNumberRepository.deleteById((financialDocumentNumber.getId())));
             financialDocumentNumberRepository.flush();
-            financialDocumentRepository.deleteById(financialDocument.getId());
+            financialDocumentRepository.deleteById(financialDocumentId);
         }
         return true;
     }
@@ -788,6 +811,12 @@ public class DefaultFinancialDocument implements FinancialDocumentService {
         FinancialPeriodStatusResponse financialPeriodStatus = financialPeriodService.getFinancialPeriodStatus(financialPeriodStatusRequest);
         if (financialPeriodStatus.getPeriodStatus() == 0L || financialPeriodStatus.getMonthStatus() == 0L) {
             throw new RuleException("دوره مالی و ماه عملیاتی سند مقصد میبایست در وضعیت باز باشند");
+        }
+        FinancialPeriodLedgerStatusRequest financialPeriodLedgerStatusRequest = new FinancialPeriodLedgerStatusRequest();
+        financialPeriodLedgerStatusRequest.setFinancialDocumentId(financialDocumentAccountDto.getId());
+        FinancialPeriodStatusResponse financialPeriodStatusResponse = getFinancialPeriodStatus(financialPeriodLedgerStatusRequest);
+        if (financialPeriodStatusResponse.getPeriodStatus() == 0L || financialPeriodStatusResponse.getMonthStatus() == 0L) {
+            throw new RuleException("دوره مالی و ماه مربوط به دفتر مالی میبایست در وضعیت باز باشند");
         }
         AtomicReference<String> message = new AtomicReference<>("");
         if (financialDocumentAccountDto.getFinancialAccountId().equals(financialDocumentAccountDto.getNewFinancialAccountId())) {
@@ -849,6 +878,12 @@ public class DefaultFinancialDocument implements FinancialDocumentService {
             FinancialPeriodStatusResponse financialPeriodStatus = financialPeriodService.getFinancialPeriodStatus(financialPeriodStatusRequest);
             if (financialPeriodStatus.getPeriodStatus() == 0L || financialPeriodStatus.getMonthStatus() == 0L) {
                 throw new RuleException("دوره مالی و ماه عملیاتی سند مبدا میبایست در وضعیت باز باشند");
+            }
+            FinancialPeriodLedgerStatusRequest financialPeriodLedgerStatusRequest = new FinancialPeriodLedgerStatusRequest();
+            financialPeriodLedgerStatusRequest.setFinancialDocumentId(financialCentricAccountDto.getId());
+            FinancialPeriodStatusResponse financialPeriodStatusResponse = getFinancialPeriodStatus(financialPeriodLedgerStatusRequest);
+            if (financialPeriodStatusResponse.getPeriodStatus() == 0L || financialPeriodStatusResponse.getMonthStatus() == 0L) {
+                throw new RuleException("دوره مالی و ماه مربوط به دفتر مالی میبایست در وضعیت باز باشند");
             }
             Long centricAccountId = financialCentricAccountDto.getCentricAccountId();
             financialDocumentItemList.forEach((FinancialDocumentItem documentItem) -> {
@@ -934,6 +969,13 @@ public class DefaultFinancialDocument implements FinancialDocumentService {
         if (financialPeriodStatus.getPeriodStatus() == 0L || financialPeriodStatus.getMonthStatus() == 0L) {
             throw new RuleException("دوره مالی و ماه عملیاتی سند مقصد میبایست در وضعیت باز باشند");
         }
+
+        FinancialPeriodLedgerStatusRequest financialPeriodLedgerStatusRequest = new FinancialPeriodLedgerStatusRequest();
+        financialPeriodLedgerStatusRequest.setFinancialDocumentId(financialCentricAccountDto.getId());
+        FinancialPeriodStatusResponse financialPeriodStatusResponse = getFinancialPeriodStatus(financialPeriodLedgerStatusRequest);
+        if (financialPeriodStatusResponse.getPeriodStatus() == 0L || financialPeriodStatusResponse.getMonthStatus() == 0L) {
+            throw new RuleException("دوره مالی و ماه مربوط به دفتر مالی میبایست در وضعیت باز باشند");
+        }
         List<FinancialDocumentItem> financialDocumentItemList =
                 financialDocumentItemRepository.findByFinancialDocumentItemIdList(financialCentricAccountDto.getFinancialDocumentItemIdList());
         financialDocumentItemList.forEach((FinancialDocumentItem documentItem) -> {
@@ -973,6 +1015,12 @@ public class DefaultFinancialDocument implements FinancialDocumentService {
         if (financialPeriodStatus.getPeriodStatus() == 0L || financialPeriodStatus.getMonthStatus() == 0L) {
             throw new RuleException("دوره مالی و ماه عملیاتی سند مقصد میبایست در وضعیت باز باشند");
         }
+        FinancialPeriodLedgerStatusRequest financialPeriodLedgerStatusRequest = new FinancialPeriodLedgerStatusRequest();
+        financialPeriodLedgerStatusRequest.setFinancialDocumentId(financialCentricAccountDto.getId());
+        FinancialPeriodStatusResponse financialPeriodStatusResponse = getFinancialPeriodStatus(financialPeriodLedgerStatusRequest);
+        if (financialPeriodStatusResponse.getPeriodStatus() == 0L || financialPeriodStatusResponse.getMonthStatus() == 0L) {
+            throw new RuleException("دوره مالی و ماه مربوط به دفتر مالی میبایست در وضعیت باز باشند");
+        }
         List<FinancialDocumentItem> financialDocumentItemList = financialDocumentItemRepository.findByFinancialDocumentItemIdList(financialCentricAccountDto.getFinancialDocumentItemIdList());
         financialDocumentItemList.forEach((FinancialDocumentItem documentItem) -> {
             documentItem.setCreditAmount(0D);
@@ -1010,6 +1058,13 @@ public class DefaultFinancialDocument implements FinancialDocumentService {
         if (financialPeriodStatus.getPeriodStatus() == 0L || financialPeriodStatus.getMonthStatus() == 0L) {
             throw new RuleException("دوره مالی و ماه عملیاتی سند مبدا میبایست در وضعیت باز باشند");
         }
+        FinancialPeriodLedgerStatusRequest financialPeriodLedgerStatusRequest = new FinancialPeriodLedgerStatusRequest();
+        financialPeriodLedgerStatusRequest.setFinancialDocumentId(financialDocumentDto.getId());
+        FinancialPeriodStatusResponse financialPeriodStatusResponse = getFinancialPeriodStatus(financialPeriodLedgerStatusRequest);
+        if (financialPeriodStatusResponse.getPeriodStatus() == 0L || financialPeriodStatusResponse.getMonthStatus() == 0L) {
+            throw new RuleException("دوره مالی و ماه مربوط به دفتر مالی میبایست در وضعیت باز باشند");
+        }
+
         entityManager.createNativeQuery("update fndc.financial_document_item  fndi_outer " +
                 "   set fndi_outer.sequence_number = " +
                 "       (select rn " +
@@ -1097,13 +1152,32 @@ public class DefaultFinancialDocument implements FinancialDocumentService {
     @Override
     @Transactional(rollbackFor = Throwable.class)
     public FinancialPeriodStatusResponse getFinancialPeriodStatus(FinancialPeriodLedgerStatusRequest financialPeriodLedgerStatusRequest) {
-        if (financialPeriodLedgerStatusRequest.getFinancialLedgerTypeId() == null && financialPeriodLedgerStatusRequest.getFinancialPeriodId() == null
+        if (financialPeriodLedgerStatusRequest.getFinancialDocumentId() == null && financialPeriodLedgerStatusRequest.getFinancialLedgerTypeId() == null
+                && financialPeriodLedgerStatusRequest.getFinancialPeriodId() == null
                 && financialPeriodLedgerStatusRequest.getDate() == null) {
             throw new RuleException("fin.financialPeriod.getStatus");
         }
+        if (financialPeriodLedgerStatusRequest.getFinancialDocumentId() != null) {
+            List<Object[]> financialPeriodListObject = financialPeriodRepository.getFinancialPeriodByFinancialDocumentId(financialPeriodLedgerStatusRequest.getFinancialDocumentId());
+            financialPeriodLedgerStatusRequest.setFinancialPeriodId(financialPeriodListObject.get(0)[0] == null ? financialPeriodLedgerStatusRequest.getFinancialPeriodId() : Long.parseLong(financialPeriodListObject.get(0)[0].toString()));
+            financialPeriodLedgerStatusRequest.setDate(financialPeriodListObject.get(0)[1] == null ? financialPeriodLedgerStatusRequest.getDate() : LocalDateTime.parse(financialPeriodListObject.get(0)[1].toString().substring(0, 10) + "T00:00"));
+            financialPeriodLedgerStatusRequest.setFinancialLedgerTypeId(financialPeriodListObject.get(0)[0] == null ? financialPeriodLedgerStatusRequest.getFinancialLedgerTypeId() : Long.parseLong(financialPeriodListObject.get(0)[0].toString()));
+
+            if (financialPeriodLedgerStatusRequest.getFinancialPeriodId() == null || financialPeriodLedgerStatusRequest.getDate() == null
+                    || financialPeriodLedgerStatusRequest.getFinancialLedgerTypeId() == null) {
+                throw new RuleException("fin.financialPeriod.getStatusCheck");
+            }
+        }
+        if (financialPeriodLedgerStatusRequest.getFinancialDocumentId() == null && financialPeriodLedgerStatusRequest.getFinancialPeriodId() == null
+                && financialPeriodLedgerStatusRequest.getDate() != null && financialPeriodLedgerStatusRequest.getFinancialLedgerTypeId() != null) {
+            FinancialPeriodRequest financialPeriodRequest = new FinancialPeriodRequest();
+            financialPeriodRequest.setOrganizationId(SecurityHelper.getCurrentUser().getOrganizationId());
+            financialPeriodRequest.setDate(financialPeriodLedgerStatusRequest.getDate());
+        }
+
         FinancialPeriodStatusResponse financialPeriodStatusResponses = new FinancialPeriodStatusResponse();
         Long periodStatus = financialPeriodRepository.findFinancialPeriodByIdAndLedgerType(financialPeriodLedgerStatusRequest.getFinancialPeriodId(), financialPeriodLedgerStatusRequest.getFinancialLedgerTypeId());
-        Long monthStatus = financialPeriodRepository.findFinancialPeriodByIdAndLedgerTypeAndDate(financialPeriodLedgerStatusRequest.getFinancialPeriodId(), financialPeriodLedgerStatusRequest.getFinancialLedgerTypeId(),DateUtil.convertStringToDate(financialPeriodLedgerStatusRequest.getDate().format(DateTimeFormatter.ofPattern("MM/dd/yyyy"))), financialPeriodLedgerStatusRequest.getDate().format(DateTimeFormatter.ofPattern("MM/dd/yyyy")));
+        Long monthStatus = financialPeriodRepository.findFinancialPeriodByIdAndLedgerTypeAndDate(financialPeriodLedgerStatusRequest.getFinancialPeriodId(), financialPeriodLedgerStatusRequest.getFinancialLedgerTypeId(), DateUtil.convertStringToDate(financialPeriodLedgerStatusRequest.getDate().format(DateTimeFormatter.ofPattern("MM/dd/yyyy"))), financialPeriodLedgerStatusRequest.getDate().format(DateTimeFormatter.ofPattern("MM/dd/yyyy")));
 
         financialPeriodStatusResponses.setPeriodStatus(periodStatus);
         financialPeriodStatusResponses.setMonthStatus(monthStatus);
